pyproject.toml
[tool.poetry]
name = "darqube-api"
version = "0.1.0"
description = ""
authors = ["Daniil Sheremet <daniil.shuhari@gmail.com>"]
readme = "README.md"
packages = [{include = "darqube_api"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
motor = "^3.0.0"
bcrypt = "^4.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
uvicorn = "^0.18.3"
python-multipart = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
